[FrameworkBundle][Validator] Remove remaining deprecations
