+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/schema/symfony-1.0.xsd
@@ -266,7 +266,6 @@
 
         <xsd:attribute name="enabled" type="xsd:boolean" />
         <xsd:attribute name="cache" type="xsd:string" />
-        <xsd:attribute name="enable-annotations" type="xsd:boolean" />
         <xsd:attribute name="enable-attributes" type="xsd:boolean" />
         <xsd:attribute name="static-method" type="xsd:boolean" />
         <xsd:attribute name="translation-domain" type="xsd:string" />
@@ -320,7 +319,6 @@
             <xsd:element name="default-context" type="metadata" minOccurs="0" maxOccurs="1" />
         </xsd:choice>
         <xsd:attribute name="enabled" type="xsd:boolean" />
-        <xsd:attribute name="enable-annotations" type="xsd:boolean" />
         <xsd:attribute name="enable-attributes" type="xsd:boolean" />
         <xsd:attribute name="name-converter" type="xsd:string" />
         <xsd:attribute name="circular-reference-handler" type="xsd:string" />
