+++ src/Symfony/Component/Validator/Mapping/Loader/AttributeLoader.php
@@ -11,6 +11,13 @@
 
 namespace Symfony\Component\Validator\Mapping\Loader;
 
+use Symfony\Component\Validator\Constraint;
+use Symfony\Component\Validator\Constraints\Callback;
+use Symfony\Component\Validator\Constraints\GroupSequence;
+use Symfony\Component\Validator\Constraints\GroupSequenceProvider;
+use Symfony\Component\Validator\Exception\MappingException;
+use Symfony\Component\Validator\Mapping\ClassMetadata;
+
 /**
  * Loads validation metadata using PHP attributes.
  *
@@ -18,10 +25,71 @@ namespace Symfony\Component\Validator\Mapping\Loader;
  * @author Alexander M. Turek <me@derrabus.de>
  * @author Alexandre Daubois <alex.daubois@gmail.com>
  */
-class AttributeLoader extends AnnotationLoader
+class AttributeLoader implements LoaderInterface
 {
-    public function __construct()
+    public function loadClassMetadata(ClassMetadata $metadata): bool
     {
-        parent::__construct(null);
+        $reflClass = $metadata->getReflectionClass();
+        $className = $reflClass->name;
+        $success = false;
+
+        foreach ($this->getAttributes($reflClass) as $constraint) {
+            if ($constraint instanceof GroupSequence) {
+                $metadata->setGroupSequence($constraint->groups);
+            } elseif ($constraint instanceof GroupSequenceProvider) {
+                $metadata->setGroupSequenceProvider(true);
+            } elseif ($constraint instanceof Constraint) {
+                $metadata->addConstraint($constraint);
+            }
+
+            $success = true;
+        }
+
+        foreach ($reflClass->getProperties() as $property) {
+            if ($property->getDeclaringClass()->name === $className) {
+                foreach ($this->getAttributes($property) as $constraint) {
+                    if ($constraint instanceof Constraint) {
+                        $metadata->addPropertyConstraint($property->name, $constraint);
+                    }
+
+                    $success = true;
+                }
+            }
+        }
+
+        foreach ($reflClass->getMethods() as $method) {
+            if ($method->getDeclaringClass()->name === $className) {
+                foreach ($this->getAttributes($method) as $constraint) {
+                    if ($constraint instanceof Callback) {
+                        $constraint->callback = $method->getName();
+
+                        $metadata->addConstraint($constraint);
+                    } elseif ($constraint instanceof Constraint) {
+                        if (preg_match('/^(get|is|has)(.+)$/i', $method->name, $matches)) {
+                            $metadata->addGetterMethodConstraint(lcfirst($matches[2]), $matches[0], $constraint);
+                        } else {
+                            throw new MappingException(sprintf('The constraint on "%s::%s()" cannot be added. Constraints can only be added on methods beginning with "get", "is" or "has".', $className, $method->name));
+                        }
+                    }
+
+                    $success = true;
+                }
+            }
+        }
+
+        return $success;
+    }
+
+    private function getAttributes(\ReflectionMethod|\ReflectionClass|\ReflectionProperty $reflection): iterable
+    {
+        foreach ($reflection->getAttributes(GroupSequence::class) as $attribute) {
+            yield $attribute->newInstance();
+        }
+        foreach ($reflection->getAttributes(GroupSequenceProvider::class) as $attribute) {
+            yield $attribute->newInstance();
+        }
+        foreach ($reflection->getAttributes(Constraint::class, \ReflectionAttribute::IS_INSTANCEOF) as $attribute) {
+            yield $attribute->newInstance();
+        }
     }
 }
