+++ src/Symfony/Component/Validator/Tests/ConstraintTest.php
@@ -245,7 +245,7 @@ class ConstraintTest extends TestCase
         $this->assertEquals('foo', $constraint->property1);
     }
 
-    public function testAnnotationSetUndefinedDefaultOption()
+    public function testAttributeSetUndefinedDefaultOption()
     {
         $this->expectException(ConstraintDefinitionException::class);
         $this->expectExceptionMessage('No default option is configured for constraint "Symfony\Component\Validator\Tests\Fixtures\ConstraintB".');
