+++ src/Symfony/Component/Validator/Tests/Constraints/CallbackValidatorTest.php
@@ -205,7 +205,7 @@ class CallbackValidatorTest extends ConstraintValidatorTestCase
         $this->assertEquals($targets, $constraint->getTargets());
     }
 
-    // Should succeed. Needed when defining constraints as annotations.
+    // Should succeed. Needed when defining constraints as attributes.
     public function testNoConstructorArguments()
     {
         $constraint = new Callback();
@@ -213,14 +213,14 @@ class CallbackValidatorTest extends ConstraintValidatorTestCase
         $this->assertSame([Constraint::CLASS_CONSTRAINT, Constraint::PROPERTY_CONSTRAINT], $constraint->getTargets());
     }
 
-    public function testAnnotationInvocationSingleValued()
+    public function testAttributeInvocationSingleValued()
     {
         $constraint = new Callback(['value' => 'validateStatic']);
 
         $this->assertEquals(new Callback('validateStatic'), $constraint);
     }
 
-    public function testAnnotationInvocationMultiValued()
+    public function testAttributeInvocationMultiValued()
     {
         $constraint = new Callback(['value' => [__CLASS__.'_Class', 'validateCallback']]);
